ajuste build

+++ src/main/java/com/bufalari/building/requestDTO/BalconyDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,6 +15,6 @@ public class BalconyDTO {
     private String railingMaterial;
     private String floorMaterial;
     private String structureType;
-    private double estimatedCostPerSquareFoot;
-    private double totalEstimatedCost;
+    private BigDecimal estimatedCostPerSquareFoot;
+    private BigDecimal totalEstimatedCost;
 }
